Enemy Set Up
Created basic scripts for Enemy. They contain 2 interfaces, IMovement and IHealth which ensure that the created Enemy script is not tightly coupled to specific movement or health, which makes the system modular and easily expandable. Created prefabs for the 3 different types of enemies and added the Enemy script and gave them values accordingly.