Enemy Spawner
First, implemented a generic object pooling system that can be reused for different type of objects.
Created a scriptable object to define spawn patterns, which can be edited in the editor easily (Currently spawning at one place -> will change to make it randomized)
Usage of event driven spawning which is more efficient that continuously running a coroutine.