Tower
So multiple scripts for the tower and it's mechanics. Created scriptable objects for spells, and each spell can derive from this class. So you can make as many custom spells you want and each spell can customized to it's own liking. Then created a Projectile class which is responsible for shooting the projectiles. This again is can be used as a base class for all the projectiles that need to be shot. Still have to implement damage functionalities. Also created a Fireball visual effect, for ease of use purposes will repurpose this for the barrage projectile as well.