Projectile Arc
Created a utility script that would create an Arc like throw for the projectile. Implemented that for the Fireball spell that would throw random spells across the floor of the level. Also added a new component in the scriptable object call RequiresTarget, so this can add another level of customization for the spells